Fill out CountyMap class
It looks like there was quite a bit added here, but since the CountyMap
class inherits from the StateMap class, there wasn't actually too much
added. A little bit of it was copied and pasted code. At a later date I
will refactor this to be in a better format (as soon as I think of how
to actually format/structure everything, I will).