Add minor functionality to geo classes
This commit included adding a hodge podge of plotting functionality to
the geo classes that I've built. Some of it revolved around the look of
the plots (making them black lines instead of random colors), and others
revolved around increased functionality (being able to pass in state or
county shapefiles, being able to pass in a specified axis, etc.). I
didn't feel that any one of these things was individually large enough
for its own commit.