Adjust plot_boundary method to read in shapefile
Previously, I was passing in boundaries that were read in
using fiona, and then parsing and individually plotting those
boundaries. The parsing that I was using was dropping some parts of the
boundaries, and it turns out I can just alternatively just use the
`.readshapefile` method available on a Basemap object.