Combine map classes into one
The three classes written to plot a USMap, StateMap, and CountyMap had a
pretty decent amount of overlapping functionality. It was
easiest/quickest upon writing them for the first time to separate them
out. In the name of best practices, I've moved them all into one class.

I've named the class USMapBuilder with the idea that other maps could be
built out for other countries. It's possible this class could be
extended to map other countries, but that'll have to come when the need
arises.